+++ educacyl-javaee/src/main/java/es/jcyl/educa/javaee/comun/modelo/MunicipioId.java
@@ -11,25 +11,25 @@ public class MunicipioId implements Serializable {
 	private static final long serialVersionUID = 1L;
 
 	@Column(name = "C_PROV_ID")
-	private Long provinciaId;
+	private Integer provinciaId;
 
 	@Column(name = "C_MUNI_ID")
-	private Long municipioId;
+	private Integer municipioId;
 
 	public MunicipioId() {
 		super();
 	}
 
-	public MunicipioId(Long municipioId, Long provinciaId) {
+	public MunicipioId(Integer municipioId, Integer provinciaId) {
 		this.municipioId = municipioId;
 		this.provinciaId = provinciaId;
 	}
 
-	public Long getMunicipioId() {
+	public Integer getMunicipioId() {
 		return municipioId;
 	}
 
-	public void setMunicipioId(Long municipioId) {
+	public void setMunicipioId(Integer municipioId) {
 		this.municipioId = municipioId;
 	}
